Implemented balance, diff response msgs and some other stuff

+++ HDSLedger/Service/src/main/java/pt/ulisboa/tecnico/hdsledger/service/models/Transaction.java
@@ -25,7 +25,7 @@ public class Transaction {
         this.nonce = newNonce;
     }
 
-    public float getAmount() {
+    public int getAmount() {
         return amount;
     }
